Base project somewhat implemented with debugging

+++ .eslintrc.js
@@ -96,7 +96,7 @@ module.exports = {
     '@typescript-eslint/prefer-for-of': 'off',
     '@typescript-eslint/prefer-function-type': 'error',
     '@typescript-eslint/prefer-namespace-keyword': 'error',
-    '@typescript-eslint/quotes': ['error', 'double'],
+    '@typescript-eslint/quotes': ['error', 'single'],
     '@typescript-eslint/semi': ['error', 'always'],
     '@typescript-eslint/triple-slash-reference': [
       'error',
